package-metadata.pl: provides/depends fix
When a package PROVIDES another name, it shouldn't output dependency
information for that provided name if a real package with that name
already exists.

For example, if a Makefile contains something like this:

  define Package/bind-delv
    $(Package/bind/Default)
    TITLE+= tools (delv)
    CONFLICTS:=bind-delv-gssapi
  endef

  define Package/bind-delv-gssapi
    $(Package/bind-gssapi/Default)
    TITLE+= tools (delv) with GSSAPI
    PROVIDES:=bind-delv
  endef

The generated tmp/.packageauxvars will look like this (note the last
line):

  Package/bind-delv/subdir = packages
  Package/bind-delv/depends = bind-libs libc
  Package/bind-delv-gssapi/subdir = packages
  Package/bind-delv-gssapi/depends = bind-libs-gssapi libc
  Package/bind-delv/depends = bind-libs-gssapi libc

Fix this by skipping dependency information for provided names which
correspond to real names.

Signed-off-by: David Härdeman <david@hardeman.nu>